remove old command system
