+++ build.gradle
@@ -153,8 +153,6 @@ allprojects {
         compileOnly 'net.kyori:adventure-text-minimessage:4.13.1'
         compileOnly 'net.kyori:adventure-platform-bukkit:4.3.0'
         compileOnly 'net.kyori:adventure-api:4.13.1'
-        compileOnly 'io.github.mqzn:mCommands-spigot:1.1.0'
-        compileOnly 'io.github.mqzn:mCommands-common:1.1.0'
         compileOnly 'it.unimi.dsi:fastutil:8.5.13'
         implementation 'com.google.guava:guava:30.1-jre'
         compileOnly fileTree(dir: 'libs', include: ['*.jar'])
@@ -235,8 +233,6 @@ dependencies {
     compileOnly 'net.kyori:adventure-text-minimessage:4.13.1'
     compileOnly 'net.kyori:adventure-platform-bukkit:4.3.0'
     compileOnly 'net.kyori:adventure-api:4.13.1'
-    compileOnly 'io.github.mqzn:mCommands-spigot:1.1.0'
-    compileOnly 'io.github.mqzn:mCommands-common:1.1.0'
     compileOnly 'it.unimi.dsi:fastutil:8.5.13'
     implementation 'com.google.guava:guava:30.1-jre'
     compileOnly fileTree(dir: 'libs', include: ['*.jar'])
