--- src/main/java/com/volmit/adapt/command/debug/CommandVerbose.java
@@ -1,37 +1,0 @@
-/*------------------------------------------------------------------------------
- -   Adapt is a Skill/Integration plugin  for Minecraft Bukkit Servers
- -   Copyright (c) 2022 Arcane Arts (Volmit Software)
- -
- -   This program is free software: you can redistribute it and/or modify
- -   it under the terms of the GNU General Public License as published by
- -   the Free Software Foundation, either version 3 of the License, or
- -   (at your option) any later version.
- -
- -   This program is distributed in the hope that it will be useful,
- -   but WITHOUT ANY WARRANTY; without even the implied warranty of
- -   MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
- -   GNU General Public License for more details.
- -
- -   You should have received a copy of the GNU General Public License
- -   along with this program.  If not, see <https://www.gnu.org/licenses/>.
- -----------------------------------------------------------------------------*/
-
-package com.volmit.adapt.command.debug;
-
-import com.volmit.adapt.AdaptConfig;
-import com.volmit.adapt.util.command.FConst;
-import io.github.mqzn.commands.annotations.base.Default;
-import io.github.mqzn.commands.annotations.base.ExecutionMeta;
-import io.github.mqzn.commands.annotations.subcommands.SubCommandInfo;
-import org.bukkit.command.CommandSender;
-
-@SubCommandInfo(name = "verbose", parent = CommandDebug.class)
-@ExecutionMeta(description = "Toggle verbose mode")
-public final class CommandVerbose {
-
-    @Default
-    public void execute(CommandSender sender) {
-        AdaptConfig.get().setVerbose(!AdaptConfig.get().isVerbose());
-        FConst.success("Verbose is now " + (AdaptConfig.get().isVerbose() ? "enabled" : "disabled")).send(sender);
-    }
-}
