+++ src/main/java/com/volmit/adapt/service/CommandSVC.java
@@ -19,7 +19,7 @@
 package com.volmit.adapt.service;
 
 import com.volmit.adapt.Adapt;
-import com.volmit.adapt.commandv2.CommandAdapt;
+import com.volmit.adapt.command.CommandAdapt;
 import com.volmit.adapt.util.AdaptService;
 import com.volmit.adapt.util.J;
 import com.volmit.adapt.util.cache.AtomicCache;
