+++ src/main/resources/plugin.yml
@@ -20,8 +20,6 @@ libraries:
   - net.kyori:adventure-platform-bukkit:4.3.0
   - net.kyori:adventure-api:4.13.1
   - net.kyori:adventure-text-minimessage:4.13.1
-  - io.github.mqzn:mCommands-spigot:1.1.0
-  - io.github.mqzn:mCommands-common:1.1.0
   - it.unimi.dsi:fastutil:8.5.13
 
 commands:
